// src/types/IMatchesFromDb.ts
export interface IMatchesFromDb {
  id: string;
  idMatch: number;
  hboMax: string;
  canalTnt: string;
  playPlus: string;
  canalRecord: string;
  cazeTv: string;
  paulistaoPlay: string;
  camisa21: string;
  futebolPaulista: string;
  canalFutura: string;
  paulistao: string;
  sportv: string;
  globo: string;
  premiere: string;
  round: number;
  teamHome: string;
  teamAway: string;
}